+++ src/types.ts
@@ -1,8 +1,5 @@
-import { BlocState } from "@bloc-state/state"
+import { State } from "@bloc-state/state"
 import { Observable } from "rxjs"
-import { BlocBase } from "./base"
-import { Bloc } from "./bloc"
-import { Cubit } from "./cubit"
 import { BlocEvent } from "./event"
 
 export interface BlocEmitter<State> {
@@ -21,14 +18,14 @@ export interface BlocEmitter<State> {
 
 export type EmitUpdaterCallback<T> = (state: T) => T
 
-export type EmitDataUpdaterCallback<T extends BlocState<any>> = (state: T) => T
+export type EmitDataUpdaterCallback<T extends State<any>> = (state: T) => T
 
-export interface Emitter<S extends BlocState<any>> extends BlocEmitter<S> {
-  (state: S): void
+export interface Emitter<S> extends BlocEmitter<S> {
+  (state: S | EmitUpdaterCallback<S>): void
   close: () => void
 }
 
-export type EventHandler<E extends BlocEvent, S extends BlocState<any>> = (
+export type EventHandler<E extends BlocEvent, S> = (
   event: InstanceType<ClassType<E>>,
   emitter: Emitter<S>,
 ) => void | Promise<void>
@@ -37,35 +34,9 @@ export interface ClassType<T = any> extends Function {
   new (...args: any[]): T
 }
 
-export type BlocStateType<T extends BlocBase<any>> = T extends Cubit<infer U>
-  ? U
-  : T extends Bloc<any, infer D>
-  ? D
-  : never
-
-export type BlocStateDataType<T> = T extends BlocState<infer U> ? U : T
-
-export type CubitSelectorConfig<State, P> = {
-  selector: (state: State) => P
-  filter?: (state: P) => boolean
-}
-
-export type BlocSelectorConfig<State extends BlocState<any>, P> = {
-  selector: (state: BlocStateDataType<State>) => P
-  filter?: (state: P) => boolean
-}
-
 export type EventMapper<Event extends BlocEvent> = (
   event: Event,
 ) => Observable<void>
-
-export type OnEventConfig<
-  Event extends BlocEvent,
-  State extends BlocState<any>,
-> = {
-  listenTo?: ClassType<State>
-  transformer?: EventTransformer<Event>
-}
 
 export type EventTransformer<Event extends BlocEvent> = (
   events$: Observable<Event>,