+++ tests/change.spec.ts
@@ -3,8 +3,9 @@ import { CounterState } from "./helpers/counter/counter.state"
 
 describe("change", () => {
   it("should create a change object", () => {
-    const current = CounterState.ready(0)
-    const next = CounterState.ready(1)
+    expect.assertions(2)
+    const current = new CounterState(0)
+    const next = current.ready(1)
     const change = new Change(current, next)
 
     expect(change.currentState).toBe(current)