+++ tests/cubit.spec.ts
@@ -12,17 +12,19 @@ describe("Cubit", () => {
   })
 
   it("should create a new Cubit instance", () => {
+    expect.assertions(1)
     expect(cubit).toBeInstanceOf(Cubit)
   })
 
-  it("should close a cubit", (done) => {
+  it("should close a cubit", async () => {
+    expect.assertions(2)
+    expect(cubit.isClosed).toBe(false)
     cubit.close()
-    cubit.state$.subscribe({
-      complete: () => done(),
-    })
+    await expect(cubit.isClosed).toBe(true)
   })
 
   it("should return new state from actions", (done) => {
+    expect.assertions(4)
     const states: number[] = []
     state$.pipe(tap((state) => states.push(state))).subscribe({
       complete: () => {
@@ -40,6 +42,7 @@ describe("Cubit", () => {
   })
 
   it("should handle async actions", (done) => {
+    expect.assertions(5)
     void (async () => {
       const states: number[] = []
       state$.pipe(tap((state) => states.push(state))).subscribe({
@@ -108,6 +111,7 @@ describe("Cubit", () => {
     })
 
     it("should be invoked when an error is thrown from BlocBase.emit", () => {
+      expect.assertions(1)
       errorBloc.triggerError()
       const [a] = errors
 
@@ -115,6 +119,7 @@ describe("Cubit", () => {
     })
 
     it("should be invoked when an error is thrown from BlocBase.onChange", () => {
+      expect.assertions(1)
       try {
         errorBloc.triggerChange()
       } catch (e) {}
@@ -126,6 +131,7 @@ describe("Cubit", () => {
   })
 
   it("should not emit values if the bloc is closed", (done) => {
+    expect.assertions(4)
     const states: number[] = []
     cubit.state$.subscribe({
       next: (state) => states.push(state),
@@ -146,106 +152,5 @@ describe("Cubit", () => {
     expect(b).toBe(2)
     expect(c).toBe(3)
     done()
-  })
-
-  describe("Cubit.select", () => {
-    type Car = {
-      brand: string
-      year: number
-    }
-
-    class CarBloc extends Cubit<Car> {
-      constructor(car: Car) {
-        super(car)
-      }
-
-      brandsFiltered$ = this.select({
-        selector: (car) => car.brand,
-        filter: (brand) => brand.length <= 4,
-      })
-
-      brand$ = this.select((car) => car.brand)
-
-      year$ = this.select({
-        selector: (car) => car.year,
-      })
-
-      updateCar(car: Partial<Car>) {
-        this.emit((previous) => ({ ...previous, ...car }))
-      }
-    }
-
-    let bloc: CarBloc
-
-    beforeEach(() => {
-      bloc = new CarBloc({ brand: "ford", year: 2021 })
-    })
-
-    it("should map cars mapped state", (done) => {
-      const brands: string[] = []
-
-      bloc.brand$.subscribe({
-        next: (brand) => brands.push(brand),
-        complete: () => {
-          const [a, b, c, d] = brands
-          expect(brands.length).toBe(4)
-          expect(a).toBe("ford")
-          expect(b).toBe("toyota")
-          expect(c).toBe("mercedes")
-          expect(d).toBe("bmw")
-          done()
-        },
-      })
-
-      bloc.updateCar({ brand: "toyota" })
-      bloc.updateCar({ brand: "mercedes" })
-      bloc.updateCar({ brand: "bmw" })
-      bloc.updateCar({ year: 2022 })
-      bloc.close()
-    })
-
-    it("should map years to selected state", (done) => {
-      const years: number[] = []
-
-      bloc.year$.subscribe({
-        next: (year) => years.push(year),
-        complete: () => {
-          const [a, b] = years
-
-          expect(years.length).toBe(2)
-          expect(a).toBe(2021)
-          expect(b).toBe(2022)
-          done()
-        },
-      })
-
-      bloc.updateCar({ brand: "toyota" })
-      bloc.updateCar({ brand: "mercedes" })
-      bloc.updateCar({ brand: "bmw" })
-      bloc.updateCar({ year: 2022 })
-      bloc.close()
-    })
-
-    it("should filter selected mapped state", (done) => {
-      const brands: string[] = []
-      bloc.brandsFiltered$.subscribe({
-        next: (brand) => brands.push(brand),
-        complete: () => {
-          const [a, b] = brands
-
-          expect(brands.length).toBe(2)
-          expect(a).toBe("ford")
-          expect(b).toBe("bmw")
-          done()
-        },
-      })
-
-      bloc.updateCar({ brand: "toyota" })
-      bloc.updateCar({ brand: "mercedes" })
-      bloc.updateCar({ brand: "bmw" })
-      bloc.updateCar({ year: 2022 })
-
-      bloc.close()
-    })
   })
 })