+++ tests/emitter.spec.ts
@@ -1,4 +1,4 @@
-import { BlocState } from "@bloc-state/state"
+import { State } from "@bloc-state/state"
 import { Observable, interval, take } from "rxjs"
 import { Bloc, BlocEvent } from "../src"
 import { delay } from "./helpers/counter/delay"
@@ -9,53 +9,41 @@ describe("emitter", () => {
   class IntervalForEachEvent extends IntervalEvent {}
   class IntervalNoEmitOnCloseEvent extends IntervalEvent {}
 
-  class IntervalState extends BlocState<number> {}
+  class IntervalState extends State<number> {}
 
   class IntervalBloc extends Bloc<IntervalEvent, IntervalState> {
     constructor(stream$: Observable<number>) {
-      super(IntervalState.init(10)) // initial state is 10 to differentiate from state emitted from a stream
+      super(new IntervalState(10)) // initial state is 10 to differentiate from state emitted from a stream
 
-      this.on(
-        IntervalNoEmitOnCloseEvent,
-        async (event, emit) => {
-          await emit.forEach(stream$, (data) => {
-            return IntervalState.ready(data)
-          })
-          emit.close()
-          await delay(1000)
-          emit(IntervalState.ready(10))
-        },
-        { listenTo: IntervalState },
-      )
+      this.on(IntervalNoEmitOnCloseEvent, async (event, emit) => {
+        await emit.forEach(stream$, (data) => {
+          return this.state.ready(data)
+        })
+        emit.close()
+        await delay(1000)
+        emit(this.state.ready(10))
+      })
 
-      this.on(
-        IntervalForEachEvent,
-        async (event, emit) => {
-          await emit.forEach(
-            stream$,
-            (data) => IntervalState.ready(data),
-            (error) => IntervalState.failed(error),
-          )
-        },
-        { listenTo: IntervalState },
-      )
+      this.on(IntervalForEachEvent, async (event, emit) => {
+        await emit.forEach(
+          stream$,
+          (data) => this.state.ready(data),
+          (error) => this.state.failed(error),
+        )
+      })
 
-      this.on(
-        IntervalOnEachEvent,
-        async (event, emit) => {
-          await emit.onEach(
-            stream$,
-            (data) => {
-              emit(IntervalState.ready(data))
-            },
-            (error) => {
-              emit(IntervalState.failed(error))
-            },
-          )
-          emit(IntervalState.loading()) // set loading to trigger completion
-        },
-        {},
-      )
+      this.on(IntervalOnEachEvent, async (event, emit) => {
+        await emit.onEach(
+          stream$,
+          (data) => {
+            emit(this.state.ready(data))
+          },
+          (error) => {
+            emit(this.state.failed(error))
+          },
+        )
+        emit(this.state.loading()) // set loading to trigger completion
+      })
     }
   }
 
@@ -69,14 +57,14 @@ describe("emitter", () => {
 
   describe("emitter.onEach", () => {
     it("should emit values from a stream", (done) => {
+      expect.assertions(1)
       const states: number[] = []
       intervalBloc.state$.subscribe({
         next: (state) => {
-          if (state.payload.hasData) {
-            states.push(state.payload.data)
-          }
-          if (state.payload.loading) {
+          if (state.status === "loading") {
             intervalBloc.close()
+          } else {
+            states.push(state.data)
           }
         },
         complete: () => {
@@ -89,6 +77,7 @@ describe("emitter", () => {
     })
 
     it("should invoke onError if an error is thrown from onEach stream", (done) => {
+      expect.assertions(4)
       const states: IntervalState[] = []
       const errorStream$ = new Observable<number>((subscriber) => {
         subscriber.next(1)
@@ -102,19 +91,18 @@ describe("emitter", () => {
 
       bloc.state$.subscribe({
         next: (state) => {
-          if (state.payload.hasData || state.payload.hasError) {
-            states.push(state)
-          }
-          if (state.payload.loading) {
+          if (state.status === "loading") {
             bloc.close()
+          } else {
+            states.push(state)
           }
         },
         complete: () => {
           const [a, b, c] = states
           expect(states.length).toBe(3)
-          expect(a.payload.data).toBe(10)
-          expect(b.payload.data).toBe(1)
-          expect(c.payload.error?.message).toBe("stream error")
+          expect(a.data).toBe(10)
+          expect(b.data).toBe(1)
+          expect(c.error?.message).toBe("stream error")
           done()
         },
       })
@@ -125,12 +113,11 @@ describe("emitter", () => {
 
   describe("emitter.forEach", () => {
     it("should emit state returned from onData", async () => {
+      expect.assertions(1)
       const states: number[] = []
       intervalBloc.state$.subscribe({
         next: (state) => {
-          if (state.payload.hasData) {
-            states.push(state.payload.data)
-          }
+          states.push(state.data)
         },
         complete: () => {
           expect(states.length).toBe(4)
@@ -143,12 +130,11 @@ describe("emitter", () => {
     }, 10000)
 
     it("should not emit after an emitter has been closed", async () => {
+      expect.assertions(1)
       const states: number[] = []
       intervalBloc.state$.subscribe({
         next: (state) => {
-          if (state.payload.hasData) {
-            states.push(state.payload.data)
-          }
+          states.push(state.data)
         },
         complete: () => {
           expect(states.length).toBe(4)
@@ -160,7 +146,8 @@ describe("emitter", () => {
       intervalBloc.close()
     }, 10000)
 
-    it("should emit state returned from onError when an error is thrown", async () => {
+    it("should emit state returned from onError when an error is thrown", (done) => {
+      expect.assertions(4)
       const states: IntervalState[] = []
       const errorStream$ = new Observable<number>((subscriber) => {
         subscriber.next(1)
@@ -174,22 +161,20 @@ describe("emitter", () => {
 
       bloc.state$.subscribe({
         next: (state) => {
-          if (state.payload.hasData || state.payload.hasError) {
-            states.push(state)
-          }
+          states.push(state)
         },
         complete: () => {
           const [a, b, c] = states
           expect(states.length).toBe(3)
-          expect(a.payload.data).toBe(10)
-          expect(b.payload.data).toBe(1)
-          expect(c.payload.error?.message).toBe("stream error")
+          expect(a.data).toBe(10)
+          expect(b.data).toBe(1)
+          expect(c.error?.message).toBe("stream error")
+          done()
         },
       })
 
       bloc.add(new IntervalForEachEvent())
-      await delay(3000)
-      bloc.close()
-    }, 10000)
+      delay(1000).then(() => bloc.close())
+    })
   })
 })