+++ tests/helpers/user/index.ts
@@ -1,4 +1,4 @@
-import { BlocState } from "@bloc-state/state"
+import { State } from "@bloc-state/state"
 import { Bloc } from "../../../src/bloc"
 import { BlocEvent } from "../../../src/event"
 
@@ -10,11 +10,17 @@ export interface User {
   age: number
 }
 
-export class UserState extends BlocState<User> {}
-
-export class UserNameChangeState extends UserState {}
-
-export class UserAgeChangedState extends UserState {}
+export class UserState extends State<User> {
+  constructor() {
+    super({
+      name: {
+        first: "",
+        last: "",
+      },
+      age: 0,
+    })
+  }
+}
 
 export class UserEvent extends BlocEvent {}
 
@@ -31,48 +37,23 @@ export class UserAgeChangedEvent extends UserEvent {
 }
 
 export class UserBloc extends Bloc<UserEvent, UserState> {
-  name$ = this.select((state) => state.name)
-
-  ageWithSelectorMethod$ = this.select((state) => state.age)
-
-  age$ = this.select({
-    selector: (state) => state.age as number,
-  })
-
-  bob$ = this.select({
-    selector: (state) => state.name.first,
-    filter: (name) => name === "bob",
-  })
-
-  firstName$ = this.select({
-    selector: (state) => state.name.first,
-  })
-
   constructor() {
-    super(
-      UserNameChangeState.init({
-        name: {
-          first: "",
-          last: "",
-        },
-        age: 0,
-      }),
-    )
+    super(new UserState())
 
-    this.on(
-      UserNameChangedEvent,
-      (event, emit) => {
-        emit(UserNameChangeState.ready({ ...this.data, name: event.name }))
-      },
-      { listenTo: UserNameChangeState },
-    )
+    this.on(UserNameChangedEvent, (event, emit) => {
+      emit((state) =>
+        state.ready((data) => {
+          data.name = event.name
+        }),
+      )
+    })
 
-    this.on(
-      UserAgeChangedEvent,
-      (event, emit) => {
-        emit(UserAgeChangedState.ready({ ...this.data, age: event.age }))
-      },
-      { listenTo: UserAgeChangedState },
-    )
+    this.on(UserAgeChangedEvent, (event, emit) => {
+      emit((state) =>
+        state.ready((data) => {
+          data.age = event.age
+        }),
+      )
+    })
   }
 }